+++ src/shared/containers/SetOrg.tsx
@@ -22,6 +22,7 @@ import {
   FlowPage,
   FlowsIndex,
   HomepageContainer,
+  HomepagePythonWizard,
   LabelsIndex,
   MembersIndex,
   MePage,
@@ -282,6 +283,14 @@ const SetOrg: FC = () => {
             <Route exact path="/orgs/:orgID" component={HomepageContainer} />
           ) : (
             <Route exact path="/orgs/:orgID" component={MePage} />
+          )}
+
+          {isFlagEnabled('firstMile') && (
+            <Route
+              exact
+              path="/orgs/:orgID/new-user-wizard/python"
+              component={HomepagePythonWizard}
+            />
           )}
 
           <Route component={NotFound} />
