+++ src/shared/containers/index.tsx
@@ -5,9 +5,6 @@ export const DataExplorerPage = lazy(() =>
 )
 
 export const MePage = lazy(() => import('src/me/containers/MePage'))
-export const HomepageContainer = lazy(() =>
-  import('src/homepageExperience/containers/HomepageContainer')
-)
 export const TasksPage = lazy(() => import('src/tasks/containers/TasksPage'))
 export const TaskPage = lazy(() => import('src/tasks/containers/TaskPage'))
 export const TaskRunsPage = lazy(() =>
@@ -101,3 +98,15 @@ export const UserAccountPage = lazy(() => import('src/accounts/AccountPage'))
 export const OrgOverlay = lazy(() => import('src/operator/OrgOverlayWrapper'))
 
 export const CheckoutPage = lazy(() => import('src/checkout/CheckoutPage'))
+
+export const HomepageContainer = lazy(() =>
+  import(
+    'src/homepageExperience/containers/HomepageContainer'
+  ).then(module => ({default: module.HomepageContainer}))
+)
+
+export const HomepagePythonWizard = lazy(() =>
+  import(
+    'src/homepageExperience/containers/HomepagePythonWizard'
+  ).then(module => ({default: module.HomepagePythonWizard}))
+)
